hotel-api-lektion-9/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
psycopg==3.1.8
python-dotenv==1.0.0
